Move test util functions
